move PostUser function to signup.js
