+++ src/main/webapp/signup.js
@@ -3,8 +3,11 @@
  * @property {string} gitHubAuthToken
  * @property {string[]} interestTags
  */
-import {gitHubAuthorizer} from './authorization.js';
-import {postUser} from './user.js';
+import { gitHubAuthorizer } from './authorization.js';
+import {
+  standardizeFetchErrors,
+  makeRelativeUrlAbsolute,
+} from './js/fetch_handler.js';
 
 const AFTER_SIGNUP_REDIRECT = '/dashboard.html';
 
@@ -16,9 +19,9 @@ function initSignupForm() {
   signupForm.addEventListener('submit', submitSignup);
 
   signupForm.elements['github-link-button'].addEventListener(
-      'click', handleGitHubLink);
+    'click', handleGitHubLink);
   gitHubAuthorizer
-      .getFirebase().auth().onAuthStateChanged(handleAuthStateChanged);
+    .getFirebase().auth().onAuthStateChanged(handleAuthStateChanged);
   // Handle auth state changed on page load in case of existing authentication.
   handleAuthStateChanged(gitHubAuthorizer.getUser());
 }
@@ -78,7 +81,7 @@ function submitSignup(e) {
       window.location.href = AFTER_SIGNUP_REDIRECT;
     }).catch((errorResponse) => {
       console.error(
-          `Error ${errorResponse.statusCode}: ${errorResponse.error}`);
+        `Error ${errorResponse.statusCode}: ${errorResponse.error}`);
       alert(errorResponse.userMessage);
     }).then(() => {
       gitHubLinkButton.disabled = false;
@@ -142,6 +145,29 @@ function createSignupBody() {
     gitHubAuthToken: gitHubAuthorizer.getToken(),
     interestTags,
   };
+}
+
+/**
+ * Posts a user to the servlet
+ * @param {URLSearchParams} userParams parameters with the users information
+ * @return {?Promise} Returns a prepared post user fetch request.
+ */
+function postUser(userParams) {
+  const fetchRequest = fetch(makeRelativeUrlAbsolute('/user'), {
+    method: 'POST',
+    headers: {
+      'Content-Type': 'application/x-www-form-urlencoded',
+    },
+    body: userParams,
+  });
+
+  const errorFormattedFetchRequest = standardizeFetchErrors(
+    fetchRequest,
+    'Failed to communicate with the server. Please try again later.',
+    'An error occcured while creating your account.' +
+    ' Please try again later.');
+
+  return errorFormattedFetchRequest;
 }
 
 initSignupForm();
