+++ src/main/webapp/user.js
@@ -14,39 +14,12 @@ function getUser(userID = null) { //eslint-disable-line
   const fetchRequest = fetch(makeRelativeUrlAbsolute(fetchURL));
 
   const errorFormattedFetchRequest = standardizeFetchErrors(
-      fetchRequest,
-      'Failed to communicate with the server. Please try again later.',
-      'An error occcured while retrieving this account.' +
+    fetchRequest,
+    'Failed to communicate with the server. Please try again later.',
+    'An error occcured while retrieving this account.' +
     ' Please try again later.');
 
-  return errorFormattedFetchRequest.then((response) => response.json())
-      .then((user) => {
-        return user;
-      });
+  return errorFormattedFetchRequest.then((response) => response.json());
 }
 
-/**
- * Posts a user to the servlet
- * @param {URLSearchParams} userParams parameters with the users information
- * @return {?Promise} Returns a prepared post user fetch request or null if the
- * fetch request could not be prepared.
- */
-function postUser(userParams) {
-  const fetchRequest = fetch(makeRelativeUrlAbsolute('/user'), {
-    method: 'POST',
-    headers: {
-      'Content-Type': 'application/x-www-form-urlencoded',
-    },
-    body: userParams,
-  });
-
-  const errorFormattedFetchRequest = standardizeFetchErrors(
-      fetchRequest,
-      'Failed to communicate with the server. Please try again later.',
-      'An error occcured while creating your account.' +
-    ' Please try again later.');
-
-  return errorFormattedFetchRequest;
-}
-
-export {getUser, postUser};
+export { getUser };
